Fully implemented public static DrangAndDrop  Class
As of now:
DragAndDrop class is designed so that any script can call it and pass the proper parameters to select an item  from a UI canvas menu and move said item from original to another (including initial slot). If the dragged item falls is not dropped off at slot it will be sent back to original slot.